reformat(structure): moved all api files to one folder except agent
